+++ apps/projex/src/app/components/home/featured/featured.model.ts
@@ -1,0 +1,7 @@
+export class FeaturedProject {
+  public title!: string;
+  //public citation!: string;
+  public id!: string;
+  public image!: string;
+  public description!: string;
+}