+++ apps/projex/src/app/components/utils/utils.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SimulationProjectUtilsModule } from '@biosimulations/simulation-project-utils';
+import { UtilsRoutingModule } from './utils-routing.module';
+
+@NgModule({
+  imports: [CommonModule, SimulationProjectUtilsModule, UtilsRoutingModule],
+})
+export class UtilsModule {}